feat: School Registration Code Refactored
